==> Code/01_data_check_and_make_new_data.ipynb <==
# Mediwhale Assignment

## 

### 1. Settings

# load packages
import numpy as np
import pandas as pd

# load data
sc = pd.read_csv('../Data/samchild.csv')
sa = pd.read_csv('../Data/samadult.csv')
fm = pd.read_csv('../Data/familyxx.csv')

### 2. data check

# extract the necessary variables
fm2 = fm[['HHX', 'FMX', 'FM_STRCP', 'FSBALANC']]

sc2 = sc[['HHX', 'FMX', 'FPX', 'SEX', 'AGE_P', 'RACERPI2',
          'REGION', 'CHGHT_TC', 'CWGHT_TC', 'BMI_SC',
           'CCONDRR6']]

sa2 = sa[['HHX', 'FMX', 'FPX', 'SEX', 'AGE_P', 'RACERPI2',
          'REGION', 'HYPEV', 'HYPMED2',
          'CHLEV', 'CHLMDNW2', 'CHDEV', 'STREV',
          'DIBEV1', 'DIBPRE2', 'DIBAGE1', 'DIBPILL1',
          'INSLN1', 'KIDWKYR', 'PREGNOW', 'PREGFLYR', 'SMKEV',
          'SMKNOW', 'SMKQTTP', 'SMKQTNO', 'ALCAMT',
          'ALC12MWK', 'AHEIGHT', 'AWEIGHTP', 
          'BMI', 'ASICPUSE', 'ASISLEEP', 'DEP_1', 'DEP_2',
          'VIGFREQW','MODFREQW', 'STRFREQW']]

# details
fm2.dtypes

sc2.dtypes

sa2.dtypes

# check NA
fm2.isna().sum()

sc2.isna().sum()

sa2.isna().sum().sort_values(ascending = False)

# describe data
fm2.describe()

sc2.describe()

sa2.describe()

### 3. Data outlier and NA values imputate

#### 1) familyxx.csv

fm2[fm2['HHX'] == 87]

fm2['FM_STRCP'].value_counts()

# 동거인이 있으면 0 없으면 1 Unknown은 np.nan
fm2['live_alone'] = 1
fm2['live_alone'][fm2['FM_STRCP'] != 11] = 0
fm2['live_alone'][fm2['FM_STRCP'] == 99] = np.nan

fm2['FSBALANC'].value_counts()

# 불균형 식사가 이루어지면 1 안 이루어지면 0 결측치는 np.nan
fm2['imbal_meal'] = 1
fm2['imbal_meal'][fm2['FSBALANC'] == 3] = 0
fm2['imbal_meal'][fm2['FSBALANC'] == 7] = np.nan
fm2['imbal_meal'][fm2['FSBALANC'] == 9] = np.nan

# 사용한 열 제거
fm3 = fm2.drop(['FM_STRCP', 'FSBALANC'], axis=1)

#### 2) samples

- age, sex, region 문제 없음

sa2['RACERPI2'].value_counts()

# 분류하기 힘든 인종은 multiple race로 포함
sa2['RACERPI2'][sa2['RACERPI2'] == 6] = 5
sc2['RACERPI2'][sc2['RACERPI2'] == 6] = 5

sa2['HYPEV'].value_counts()

sa2['HYPMED2'].value_counts()

# 고혈압 2: 약물치료까지 병행, 1: 진단 받은적 있음 0: 진단도 받은적 없음

### 조건과 선택할 배열들 정의
condlist = [(sa2['HYPEV'] == 1) & (sa2['HYPMED2'] == 1),
            (sa2['HYPEV'] == 1),
            (sa2['HYPEV'] == 2)]
choicelist = [2, 1, 0]

### 조건에 따라 배열 선택
sa2['hypert'] = np.select(condlist, choicelist, default=np.nan)

# 콜레스테롤 2: 약물치료까지 병행, 1: 진단 받은적 있음 0: 진단도 받은적 없음

### 조건과 선택할 배열들 정의
condlist = [(sa2['CHLEV'] == 1) & (sa2['CHLMDNW2'] == 1),
            (sa2['CHLEV'] == 1),
            (sa2['CHLEV'] == 2)]
choicelist = [2, 1, 0]

### 조건에 따라 배열 선택
sa2['chol'] = np.select(condlist, choicelist, default=np.nan)

# 관상동맥질환 1: 진단 받은적 있음 0: 진단도 받은적 없음

### 조건과 선택할 배열들 정의
condlist = [(sa2['CHDEV'] == 1),
            (sa2['CHDEV'] == 2)]
choicelist = [1, 0]

### 조건에 따라 배열 선택
sa2['chdev'] = np.select(condlist, choicelist, default=np.nan)

# 뇌졸중 1: 진단 받은적 있음 0: 진단도 받은적 없음

### 조건과 선택할 배열들 정의
condlist = [(sa2['STREV'] == 1),
            (sa2['STREV'] == 2)]
choicelist = [1, 0]

### 조건에 따라 배열 선택
sa2['strev'] = np.select(condlist, choicelist, default=np.nan)

# 우울증 2: 약물치료까지 병행, 1: 진단 받은적 있음 0: 진단도 받은적 없음

### 조건과 선택할 배열들 정의
condlist = [(sa2['DEP_2'] == 1),
            (sa2['DEP_1'].isin([1, 2, 3])),
            (sa2['CHLEV'].isin([4, 5]))]
choicelist = [2, 1, 0]

### 조건에 따라 배열 선택
sa2['depress'] = np.select(condlist, choicelist, default=np.nan)

### 임신상태 1: 최근까지 임신이었거나 현재 임신중 0: 아님
### 임신 변수 같은 경우 해당되는 사람 이외에는 아닌 경우로 취급

### 조건에 따라 선택
sa2['preg'] = np.where((sa2['PREGNOW'] == 1) | (sa2['PREGFLYR'] ==1), 1, 0)

sa2['KIDWKYR'].value_counts()

# 신장 약화가 이루어지면 1 안 이루어지면 0 결측치는 np.nan
sa2['kidney'] = 1
sa2['kidney'][sa2['KIDWKYR'] == 2] = 0
sa2['kidney'][sa2['KIDWKYR'] == 7] = np.nan
sa2['kidney'][sa2['KIDWKYR'] == 9] = np.nan

# 당뇨진단 2: 당뇨, 1: 당뇨전단계 0: 정상

### 조건과 선택할 배열들 정의
condlist = [(sa2['DIBEV1'] == 1),
            (sa2['DIBEV1'] == 3) | (sa2['DIBPRE2'] == 1),
            (sa2['DIBEV1'] == 2) | (sa2['DIBPRE2'] == 2)]
choicelist = [2, 1, 0]

### 조건에 따라 배열 선택
sa2['dm'] = np.select(condlist, choicelist, default=np.nan)

# 1형 당뇨와 2형 당뇨 구분해보기

### 조건과 선택할 배열들 정의
condlist = []
choicelist = [2, 1, 0]

### 조건에 따라 배열 선택
sa2['dm'] = np.where((sa2['dm'] == 2) & (sa2['DIBAGE1'] < 18), 3, sa2['dm'])
sa2['heredity'] = np.where((sa2['dm'] == 3), 1, 0)


# 유전 가정만 따로 빼서 보관하기
here = sa2[['HHX', 'FMX', 'heredity']].groupby(['HHX', 'FMX']).max().reset_index()
here['heredity'].value_counts()

sa2['SMKEV'].value_counts()

# 흡연 여부 1: 기 2: 아니

### 조건과 선택할 배열들 정의
condlist = [(sa2['SMKEV'] == 1) & (sa2['SMKNOW'] == 3) & (sa2['SMKQTTP'] == 4) &
            (sa2['SMKQTNO'] >= 5) & (sa2['SMKQTNO'] <= 96),
            (sa2['SMKEV'] == 1),
            (sa2['SMKEV'] == 2)]
choicelist = [0, 1, 0]

### 조건에 따라 배열 선택
sa2['smoke'] = np.select(condlist, choicelist, default=np.nan)

# 음주 여부 1: 음주 기준치 이상 2: 이하

### 조건과 선택할 배열들 정의
condlist = [((sa2['ALC12MWK'] == 95) | ((sa2['ALC12MWK'] >= 0) & (sa2['ALC12MWK'] <= 2))) &
            (sa2['ALCAMT'] >= 5),
            ((sa2['ALC12MWK'] == 95) | ((sa2['ALC12MWK'] >= 0) & (sa2['ALC12MWK'] <= 2))),
            (sa2['ALC12MWK'] >= 3) & (sa2['ALC12MWK'] <= 7)]
choicelist = [1, 0, 1]

### 조건에 따라 배열 선택
sa2['drink'] = np.select(condlist, choicelist, default=0)

sa2['AHEIGHT'].value_counts()

# 키 inch를 cm로 바꾸기
sa2['height'] = np.where(sa2['AHEIGHT'].isin([96, 97, 98, 99]), np.nan, round(sa2['AHEIGHT']*2.54, 1))
sc2['height'] = np.where(sc2['CHGHT_TC'].isin([96, 97, 98, 99]), np.nan, round(sc2['CHGHT_TC']*2.54, 1))

# 몸무게 pound를 kg으로 바꾸기
sa2['weight'] = np.where(sa2['AWEIGHTP'].isin([996, 997, 998, 999]), np.nan, round(sa2['AWEIGHTP']*0.453592, 1))
sc2['weight'] = np.where(sc2['CWGHT_TC'].isin([996, 997, 998, 999]), np.nan, round(sc2['CWGHT_TC']*0.453592, 1))

# bmi 계산하기
sa2['bmi'] = sa2['weight']/(sa2['height']**2)*100**2
sc2['bmi'] = sc2['weight']/(sc2['height']**2)*100**2

sa2['ASICPUSE'].value_counts()

# 컴퓨터 사용정도: 결측치는 다 빼기

### 조건에 따라 배열 선택
sa2['comp_use'] = np.where(sa2['ASICPUSE'].isin([7, 8, 9]), np.nan, sa2['ASICPUSE'])

# 수면 시간(시간 단위): 결측치는 다 빼기

### 조건에 따라 배열 선택
sa2['sleep_hour'] = np.where(sa2['ASISLEEP'].isin([97, 98, 99]), np.nan, sa2['ASISLEEP'])

sc2['CCONDRR6'].value_counts()

# 결측치 2건 단순하게 NO로 반영하기
sc2['dm'] = np.where(sc2['CCONDRR6'].isin([7, 9]), 2, sc2['CCONDRR6'])

# 운동 여부 1: 주2회 이상 2: 안함

### 조건과 선택할 배열들 정의
condlist = [((sa2['VIGFREQW'] >= 2) & (sa2['VIGFREQW'] < 30)),
    (sa2['VIGFREQW'] == 95) | ((sa2['VIGFREQW'] >= 0) & (sa2['VIGFREQW'] <= 1)),
    ((sa2['MODFREQW'] >= 2) & (sa2['MODFREQW'] < 30)),
    (sa2['MODFREQW'] == 95) | ((sa2['MODFREQW'] >= 0) & (sa2['MODFREQW'] <= 1)),
    ((sa2['STRFREQW'] >= 2) & (sa2['STRFREQW'] < 30)),
    (sa2['STRFREQW'] == 95) | ((sa2['STRFREQW'] >= 0) & (sa2['STRFREQW'] <= 1))
    ]
choicelist = [1, 0, 1, 0, 1, 0]

### 조건에 따라 배열 선택
sa2['exercise'] = np.select(condlist, choicelist, default=np.nan)

###  3. data merge

# need variables
fm3 = fm2[['HHX', 'FMX', 'live_alone', 'imbal_meal']]
sc3 = sc2[['HHX', 'FMX', 'FPX', 'SEX', 'AGE_P', 'REGION',
            'RACERPI2', 'height', 'weight', 'bmi', 'dm']]
sa3 = sa2[['HHX', 'FMX', 'FPX', 'SEX', 'AGE_P', 'REGION',
           'RACERPI2', 'height', 'weight', 'bmi', 'hypert',
           'chol', 'chdev', 'strev', 'depress', 'preg',
           'kidney', 'dm', 'smoke', 'drink', 'height', 'weight',
           'bmi', 'comp_use', 'sleep_hour', 'exercise']]

# child 구성하기
tmp = pd.merge(sc3, fm3, on = ['HHX', 'FMX'], how = 'left')
tmp2 = pd.merge(tmp, here, on = ['HHX', 'FMX'], how = 'left')

# 당뇨병 1: 진단을 받았거나, 유전적으로 존재하는 경우, 0: 그외

### 조건과 선택할 배열들 정의
condlist = [(tmp2['dm'] == 1),
            (tmp2['heredity'] == 1),
            (tmp2['dm'] == 0) | (tmp2['heredity'] == 0)]
choicelist = [1, 1, 0]

### 조건에 따라 배열 선택
tmp2['dm2'] = np.select(condlist, choicelist, default=np.nan)

tmp2.drop(['dm', 'heredity'], axis = 1).dropna()['dm2'].value_counts()

tmp2['heredity'].value_counts()

tmp2_dropna['heredity'].value_counts()

- child의 정보에서 당뇨를 가려낼 dm변수와 당뇨 유전 가능성인 heredity의 변수가 현저히 적음 -> 모델링하기 부적합하다 판단됨.

# adult 구성하기
tp = pd.merge(sa3, fm3, on = ['HHX', 'FMX'], how = 'left')

# 1형 당뇨 환자 제거하기
adult = tp[tp['dm'] != 3]

# 결측치 있는 행 제거하기
a = adult.dropna()

# 당뇨병을 앓지 않은 사람이 많지만 대체적으로 정보가 많은 편
# 깔끔하게 만들어둔 데이터를 활용하기로 한다.
a['dm'].value_counts()

# 데이터 저장
a.to_csv("../Data/adult_use_data.csv", index = False)
