==> Code/02_data_train_with_boostmodel.ipynb <==
# Mediwhale Assignment

## 

### 1. Settings

# load packages
import numpy as np
import pandas as pd
import pickle

# model
from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix
from sklearn.metrics import classification_report
from xgboost import XGBClassifier
from lightgbm import LGBMClassifier

# load data
data = pd.read_csv('../Data/adult_use_data.csv')

# drop HHX, FMX, FPX
df = data.drop(['HHX', 'FMX', 'FPX', 'height.1', 'weight.1', 'bmi.1'], axis = 1)

df.describe()

df.dtypes

### 2. Split Train Test

# 모델의 robust를 확보하기 위해 층화추출로 분리
X = df.drop("dm",axis=1)
y = df["dm"]
X_train, X_test, y_train, y_test =train_test_split(X,
                                                    y,
                                                    test_size=0.3,
                                                    shuffle=True,
                                                    stratify=y,
                                                    random_state= 42)

### 3. xgb, lgbm

# XGB 
xgb = XGBClassifier(random_state=42)

# XGB 적합
xgb.fit(X_train, y_train)

# Confusion matrix
xgb_pred = xgb.predict(X_test)
print('\n','Confusion Matrix - Test:','\n',confusion_matrix(y_test, xgb_pred))

print(classification_report(y_test, xgb_pred))

# LGBM
lgbm = LGBMClassifier(force_row_wise=True, 
                      random_state=42)

# lgbm 적합
lgbm.fit(X_train, y_train)

# Confusion matrix
lgbm_pred = lgbm.predict(X_test)
print('\n','Confusion Matrix - Test:','\n',confusion_matrix(y_test, lgbm_pred))

print(classification_report(y_test, lgbm_pred))

# save the model
with open('../Data/model_xgb.pickle','wb') as fw:
    pickle.dump(xgb, fw)

with open('../Data/model_lgbm.pickle','wb') as fw:
    pickle.dump(lgbm, fw)
